Create metamaskrobberyfundsgoneREADME.md
My tokens are not in my Metamask wallet.  I went through the whole gamut of trying to get it back on my MM account to no avail.  I went on Etherscan.io and it appears that some transactions were made that I did not initiate.  What is the purpose of so-called "blockchain" that is supposed to be private and secure when hackers are winning at this game.  There is no one to call or help retrieve monies.  At least with a bank you can sleep at night knowing your money is there.  Metamask sucks!!